Improve ThemeProvider
