+++ apps/storefront/src/app/[locale]/(main)/collections/[slug]/page.tsx
@@ -126,7 +126,7 @@ export default async function Page(props: PageProps) {
           />
         ) : null}
       </div>
-      <div className="prose dark:prose-headings:text-stone-200 dark:tw-prose-body:text-stone-200 dark:prose-p:text-stone-200 dark:prose-li:text-stone-200 dark:prose-blockquote:text-stone-200 dark:prose-li:marker:text-stone-200 max-w-none">
+      <div>
         <RichText contentData={collection?.description} />
       </div>
