+++ apps/storefront/src/lib/scripts/theme-preload-script.ts
@@ -1,0 +1,12 @@
+export const themePreloadScript = `
+try {
+  const theme = localStorage.getItem("theme");
+  const systemPrefersDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
+  const isDark = theme === "dark" || (!theme && systemPrefersDark);
+  if (isDark) {
+    document.documentElement.classList.add("dark");
+  } else {
+    document.documentElement.classList.remove("dark");
+  }
+} catch(e) {}
+`;
