+++ packages/ui/src/components/rich-text/editorjs-renderer.tsx
@@ -21,7 +21,7 @@ export function EditorjsRenderer({
     <article
       className={cn(
         { prose: !disableProse },
-        "max-w-none [&>*]:mt-0",
+        "prose dark:prose-headings:text-stone-200 dark:tw-prose-body:text-stone-200 dark:prose-p:text-stone-200 dark:prose-li:text-stone-200 dark:prose-blockquote:text-stone-200 dark:prose-li:marker:text-stone-200 max-w-none [&>*]:mt-0",
         className,
       )}
       {...props}
